+++ lib/librarian/puppet/source/local.rb
@@ -20,7 +20,7 @@ module Librarian
           end
 
           install_path = environment.install_path.join(organization_name(name))
-          if install_path.exist?
+          if install_path.exist? && rsync? != true
             debug { "Deleting #{relative_path_to(install_path)}" }
             install_path.rmtree
           end
